♻️ FIX: 달력 렌더링 최적화 및 추상화
